feat(Badge): migrate to Tailwind
