--- packages/orbit-components/src/tmp-tailwind/Badge/index.js.flow
@@ -1,36 +1,0 @@
-// @flow
-/*
-  DOCUMENTATION: https://orbit.kiwi/components/badge/
-*/
-import * as React from "react";
-
-import type { Carrier } from "../../CarrierLogo/index.js.flow";
-import type { Globals } from "../../common/common.js.flow";
-
-export type Type =
-  | "neutral"
-  | "dark"
-  | "info"
-  | "success"
-  | "warning"
-  | "critical"
-  | "white"
-  | "infoSubtle"
-  | "criticalSubtle"
-  | "successSubtle"
-  | "warningSubtle"
-  | "bundleBasic"
-  | "bundleMedium"
-  | "bundleTop";
-
-export type Props = {|
-  +children?: React.Node,
-  +carriers?: Carrier[],
-  +border?: boolean,
-  +type?: Type,
-  +icon?: React.Node,
-  +ariaLabel?: string,
-  ...Globals,
-|};
-
-declare export default React.ComponentType<Props>;
